+++ com.unity.render-pipelines.universal/Runtime/Passes/PostProcessPass.cs
@@ -86,6 +86,18 @@ namespace UnityEngine.Rendering.Universal.Internal
         // Renderer is using swapbuffer system
         bool m_UseSwapBuffer;
 
+        // RTHandle alias for _TempTarget
+        RTHandle m_TempTarget = RTHandles.Alloc(ShaderConstants._TempTarget);
+
+        // RTHandle alias for _TempTarget2
+        RTHandle m_TempTarget2 = RTHandles.Alloc(ShaderConstants._TempTarget2);
+
+        // RTHandle used as a temporary target when operations need to be performed before image scaling
+        RTHandle m_ScalingSetupTarget;
+
+        // True if there are passes that will run after post processing logic and before final post
+        bool m_hasExternalPostPasses;
+
         Material m_BlitMaterial;
 
         public PostProcessPass(RenderPassEvent evt, PostProcessData data, Material blitMaterial)
@@ -141,7 +153,7 @@ namespace UnityEngine.Rendering.Universal.Internal
 
         public void Cleanup() => m_Materials.Cleanup();
 
-        public void Setup(in RenderTextureDescriptor baseDescriptor, in RenderTargetHandle source, bool resolveToScreen, in RenderTargetHandle depth, in RenderTargetHandle internalLut, bool hasFinalPass, bool enableSRGBConversion)
+        public void Setup(in RenderTextureDescriptor baseDescriptor, in RenderTargetHandle source, bool resolveToScreen, in RenderTargetHandle depth, in RenderTargetHandle internalLut, bool hasFinalPass, bool enableSRGBConversion, bool hasExternalPostPasses)
         {
             m_Descriptor = baseDescriptor;
             m_Descriptor.useMipMap = false;
@@ -155,9 +167,10 @@ namespace UnityEngine.Rendering.Universal.Internal
             m_ResolveToScreen = resolveToScreen;
             m_Destination = RenderTargetHandle.CameraTarget;
             m_UseSwapBuffer = true;
+            m_hasExternalPostPasses = hasExternalPostPasses;
         }
 
-        public void Setup(in RenderTextureDescriptor baseDescriptor, in RenderTargetHandle source, RenderTargetHandle destination, in RenderTargetHandle depth, in RenderTargetHandle internalLut, bool hasFinalPass, bool enableSRGBConversion)
+        public void Setup(in RenderTextureDescriptor baseDescriptor, in RenderTargetHandle source, RenderTargetHandle destination, in RenderTargetHandle depth, in RenderTargetHandle internalLut, bool hasFinalPass, bool enableSRGBConversion, bool hasExternalPostPasses)
         {
             m_Descriptor = baseDescriptor;
             m_Descriptor.useMipMap = false;
@@ -170,9 +183,10 @@ namespace UnityEngine.Rendering.Universal.Internal
             m_HasFinalPass = hasFinalPass;
             m_EnableSRGBConversionIfNeeded = enableSRGBConversion;
             m_UseSwapBuffer = false;
+            m_hasExternalPostPasses = hasExternalPostPasses;
         }
 
-        public void SetupFinalPass(in RenderTargetHandle source, bool useSwapBuffer = false)
+        public void SetupFinalPass(in RenderTargetHandle source, bool useSwapBuffer = false, bool hasExternalPostPasses = true)
         {
             m_Source = source.id;
             m_Destination = RenderTargetHandle.CameraTarget;
@@ -180,6 +194,7 @@ namespace UnityEngine.Rendering.Universal.Internal
             m_HasFinalPass = false;
             m_EnableSRGBConversionIfNeeded = true;
             m_UseSwapBuffer = useSwapBuffer;
+            m_hasExternalPostPasses = hasExternalPostPasses;
         }
 
         /// <inheritdoc/>
@@ -519,6 +534,22 @@ namespace UnityEngine.Rendering.Universal.Internal
 
                 if (RequireSRGBConversionBlitToBackBuffer(cameraData))
                     m_Materials.uber.EnableKeyword(ShaderKeywordStrings.LinearToSRGBConversion);
+
+                // When we're running FSR upscaling and there's no passes after this (including the FXAA pass), we can safely perform color conversion as part of uber post
+
+                // When FSR is active, we're required to provide it with input in a perceptual color space. Ideally, we can just do the color conversion as part of UberPost
+                // since FSR will *usually* be executed right after it. Unfortunately, there are a couple of situations where this is not true:
+                // 1. It's possible for users to add their own passes between UberPost and FinalPost. When user passes are present, we're unable to perform the conversion
+                //    here since it'd change the color space that the passes operate in which could lead to incorrect results.
+                // 2. When FXAA is enabled with FSR, FXAA is moved to an earlier pass to ensure that FSR sees fully anti-aliased input. The moved FXAA pass sits between
+                //    UberPost and FSR so we can no longer perform color conversion here without affecting other passes.
+                bool doEarlyFsrColorConversion = (!m_hasExternalPostPasses &&
+                                                  (((cameraData.imageScalingMode == ImageScalingMode.Upscaling) && (cameraData.upscalingFilter == ImageUpscalingFilter.FSR)) &&
+                                                   (cameraData.antialiasing != AntialiasingMode.FastApproximateAntialiasing)));
+                if (doEarlyFsrColorConversion)
+                {
+                    m_Materials.uber.EnableKeyword(ShaderKeywordStrings.Gamma20);
+                }
 
                 if (m_UseFastSRGBLinearConversion)
                 {
@@ -1395,6 +1426,16 @@ namespace UnityEngine.Rendering.Universal.Internal
 
             if (cameraData.imageScalingMode != ImageScalingMode.None)
             {
+                // FSR is only considered "enabled" when we're performing upscaling. (downscaling uses a linear filter unconditionally)
+                bool isFsrEnabled = ((cameraData.imageScalingMode == ImageScalingMode.Upscaling) && (cameraData.upscalingFilter == ImageUpscalingFilter.FSR));
+
+                bool doLateFsrColorConversion = (isFsrEnabled && (isFxaaEnabled || m_hasExternalPostPasses));
+
+                // When FXAA is enabled in scaled renders, we execute it in a separate blit since it's not designed to be used in
+                // situations where the input and output resolutions do not match.
+                // When FSR is active and we didn't perform color conversion earlier, we do it now as part of the setup blit.
+                bool isSetupRequired = (isFxaaEnabled || doLateFsrColorConversion);
+
                 // Make sure to remove any MSAA and attached depth buffers from the temporary render targets
                 var tempRtDesc = cameraData.cameraTargetDescriptor;
                 tempRtDesc.msaaSamples = 1;
@@ -1402,17 +1443,26 @@ namespace UnityEngine.Rendering.Universal.Internal
 
                 m_Materials.scalingSetup.shaderKeywords = null;
 
-                // When FXAA is enabled in scaled renders, we execute it in a separate blit since it's not designed to be used in
-                // situations where the input and output resolutions do not match.
-                if (isFxaaEnabled)
+                var sourceRtId = m_Source;
+
+                if (isSetupRequired)
                 {
-                    m_Materials.scalingSetup.EnableKeyword(ShaderKeywordStrings.Fxaa);
+                    if (isFxaaEnabled)
+                    {
+                        m_Materials.scalingSetup.EnableKeyword(ShaderKeywordStrings.Fxaa);
+                    }
+
+                    if (doLateFsrColorConversion)
+                    {
+                        m_Materials.scalingSetup.EnableKeyword(ShaderKeywordStrings.Gamma20);
+                    }
 
                     cmd.GetTemporaryRT(ShaderConstants._ScalingSetupTexture, tempRtDesc, FilterMode.Point);
-
                     Blit(cmd, m_Source, ShaderConstants._ScalingSetupTexture, m_Materials.scalingSetup);
 
                     cmd.SetGlobalTexture(ShaderPropertyId.sourceTex, ShaderConstants._ScalingSetupTexture);
+
+                    sourceRtId = ShaderConstants._ScalingSetupTexture;
                 }
 
                 switch (cameraData.imageScalingMode)
@@ -1420,6 +1470,9 @@ namespace UnityEngine.Rendering.Universal.Internal
                     case ImageScalingMode.Upscaling:
                     {
                         // In the upscaling case, set material keywords based on the selected upscaling filter
+                        // Note: If FSR is enabled, we go down this path regardless of the current render scale. We do this because
+                        //       FSR still provides visual benefits at 100% scale. This will also make the transition between 99% and 100%
+                        //       scale less obvious for cases where FSR is used with dynamic resolution scaling.
                         switch (cameraData.upscalingFilter)
                         {
                             case ImageUpscalingFilter.Point:
@@ -1431,6 +1484,41 @@ namespace UnityEngine.Rendering.Universal.Internal
                             case ImageUpscalingFilter.Linear:
                             {
                                 // Do nothing as linear is the default filter in the shader
+                                break;
+                            }
+
+                            case ImageUpscalingFilter.FSR:
+                            {
+                                m_Materials.easu.shaderKeywords = null;
+
+                                var upscaleRtDesc = tempRtDesc;
+                                upscaleRtDesc.width = cameraData.pixelWidth;
+                                upscaleRtDesc.height = cameraData.pixelHeight;
+
+                                // EASU
+                                RenderingUtils.ReAllocateIfNeeded(ref m_UpscaledTarget, upscaleRtDesc, FilterMode.Point, TextureWrapMode.Clamp, name: "_UpscaledTexture");
+                                var fsrInputSize = new Vector2(cameraData.cameraTargetDescriptor.width, cameraData.cameraTargetDescriptor.height);
+                                var fsrOutputSize = new Vector2(cameraData.pixelWidth, cameraData.pixelHeight);
+                                FSRUtils.SetEasuConstants(cmd, fsrInputSize, fsrInputSize, fsrOutputSize);
+
+                                Blit(cmd, sourceRtId, m_UpscaledTarget, m_Materials.easu);
+
+                                // RCAS
+                                // Use the override value if it's available, otherwise use the default.
+                                float sharpness = cameraData.fsrOverrideSharpness ? cameraData.fsrSharpness : FSRUtils.kDefaultSharpnessLinear;
+
+                                // Set up the parameters for the RCAS pass unless the sharpness value indicates that it wont have any effect.
+                                if (cameraData.fsrSharpness > 0.0f)
+                                {
+                                    // RCAS is performed during the final post blit, but we set up the parameters here for better logical grouping.
+                                    material.EnableKeyword(ShaderKeywordStrings.Rcas);
+                                    FSRUtils.SetRcasConstantsLinear(cmd, sharpness);
+                                }
+
+                                // Update the source texture for the next operation
+                                cmd.SetGlobalTexture(ShaderPropertyId.sourceTex, m_UpscaledTarget);
+                                PostProcessUtils.SetSourceSize(cmd, upscaleRtDesc);
+
                                 break;
                             }
                         }
@@ -1505,6 +1593,7 @@ namespace UnityEngine.Rendering.Universal.Internal
             public readonly Material paniniProjection;
             public readonly Material bloom;
             public readonly Material scalingSetup;
+            public readonly Material easu;
             public readonly Material uber;
             public readonly Material finalPass;
             public readonly Material lensFlareDataDriven;
@@ -1519,6 +1608,7 @@ namespace UnityEngine.Rendering.Universal.Internal
                 paniniProjection = Load(data.shaders.paniniProjectionPS);
                 bloom = Load(data.shaders.bloomPS);
                 scalingSetup = Load(data.shaders.scalingSetupPS);
+                easu = Load(data.shaders.easuPS);
                 uber = Load(data.shaders.uberPostPS);
                 finalPass = Load(data.shaders.finalPostPassPS);
                 lensFlareDataDriven = Load(data.shaders.LensFlareDataDrivenPS);
@@ -1549,6 +1639,7 @@ namespace UnityEngine.Rendering.Universal.Internal
                 CoreUtils.Destroy(paniniProjection);
                 CoreUtils.Destroy(bloom);
                 CoreUtils.Destroy(scalingSetup);
+                CoreUtils.Destroy(easu);
                 CoreUtils.Destroy(uber);
                 CoreUtils.Destroy(finalPass);
             }